Calibration and how this works makes so much more sen
To get a good reading with the calibrated spectrum figure I need to do an additional step. How I get a calibration file is literally by running an observation, and using that dat file as the calibration file for ploting. With that done, I can use that same method to take normal observations-- pointed at the galaxy or wherever I want. The devs of the library probably did this so they didnt have to write another function which would do the same exact thing as the observation function.